Replaced the array error_codes with a method call
Added a new method "wait" which will do the write event wrappers around "chan event"
to ensure the event fires off, and the script clears

Eliminated the "output" method from reply

Added handling for 3xx style redirects

Added hooks for servers to cache results. They are blank for the tcllib implementation.

Added the concept of TTL (time to live) as a property that can be populated for URLS
to indicate caches both internal to the webserver and proxies along the way to the 
client how fresh a page needs to be.

Corrected several HTTP codes that were copied from tclhttpd, but were in fact wrong.

Added sizes to every call to chan copy

Reimplemented proxies, SCGI, and CGI. They now follow a common template.
SCGI and CGI are both forms of proxy.

Added the ability to turn on or off debuggining in the test script.